requirements.txt
requests==2.31.0
python-gitlab==4.2.0
python-dotenv==1.0.0
pytest==8.2.2
